라우터 기반 게시판 앱 로그인해야만 글수정/삭제 가능 구현

+++ routers/board.py
@@ -89,7 +89,14 @@ async def board_detail(request: Request, bdno: int):
 
 # 게시글 삭제하기
 @router.post("/{bdno}/delete")
-async def board_delete(bdno: int):
+async def board_delete(request:Request, bdno: int):
+    # 이미 로그인이 안되었다면 /member/login으로 이동
+    if request.session.get("user") is None:
+        return RedirectResponse(url="/member/login", status_code=303)
+
+    # 자신이 쓴글이 아니면 /board/list로 이동
+
+
     async with aiosqlite.connect(BoradDB_NAME) as db:
         await db.execute("DELETE FROM board WHERE bdno = ?", (bdno,))
         await db.commit()
@@ -100,6 +107,13 @@ async def board_delete(bdno: int):
 # 게시글 수정하기 폼
 @router.get("/{bdno}/edit", response_class=HTMLResponse)
 async def board_edit_form(request: Request, bdno: int):
+    # 이미 로그인이 안되었다면 /member/login으로 이동
+    if request.session.get("user") is None:
+        return RedirectResponse(url="/member/login", status_code=303)
+
+    # 자신이 쓴글이 아니면 /board/list로 이동
+
+
     async with aiosqlite.connect(BoradDB_NAME) as db:
         async with db.execute("SELECT * FROM board WHERE bdno = ?", (bdno,)) as cur:
             result = await cur.fetchone()
@@ -123,7 +137,14 @@ async def board_edit_form(request: Request, bdno: int):
 
 # 게시글 수정하기 처리
 @router.post("/{bdno}/edit")
-async def board_edit(bdno: int, title: str = Form(...), contents: str = Form(...)):
+async def board_edit(request: Request, bdno: int, title: str = Form(...), contents: str = Form(...)):
+    # 이미 로그인이 안되었다면 /member/login으로 이동
+    if request.session.get("user") is None:
+        return RedirectResponse(url="/member/login", status_code=303)
+
+    # 자신이 쓴글이 아니면 /board/list로 이동
+
+
     async with aiosqlite.connect(BoradDB_NAME) as db:
         await db.execute("UPDATE board SET title = ?, contents = ? WHERE bdno = ?",
                          (title, contents, bdno))
